requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.5
django-cors-headers==2.5.2
django-debug-toolbar==1.11
django-easy-pdf==0.1.1
django-extra-views==0.12.0
django-filter==2.1.0
django-smart-selects==1.5.4
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
Markdown==3.0.1
mysqlclient==1.4.1
oauthlib==3.0.1
Pillow==5.4.1
psycopg2-binary==2.8.1
PyJWT==1.7.1
PyPDF2==1.26.0
python3-openid==3.1.0
pytz==2018.9
reportlab==3.5.21
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.0.0
sqlparse==0.2.4
urllib3==1.24.1
webencodings==0.5.1
whitenoise==4.1.2
xhtml2pdf==0.2.3
